refactor(rust): move observability into iii-helpers observability module
