+++ sdk/packages/rust/helpers/Cargo.toml
@@ -19,3 +19,18 @@ path = "src/lib.rs"
 serde = { version = "1", features = ["derive"] }
 serde_json = { version = "1", features = ["unbounded_depth"] }
 schemars = "0.8"
+tokio = { version = "1", features = ["macros", "rt-multi-thread", "sync", "time"] }
+tokio-tungstenite = { version = "0.28", features = ["rustls-tls-native-roots"] }
+futures-util = "0.3"
+tracing = "0.1"
+uuid = { version = "1", features = ["v4", "serde"] }
+opentelemetry = { version = "0.31" }
+opentelemetry_sdk = { version = "0.31", features = ["rt-tokio", "logs", "metrics", "trace"] }
+sysinfo = { version = "0.38" }
+reqwest = { version = "0.12", default-features = false, features = ["json", "rustls-tls"] }
+opentelemetry-http = { version = "0.31", features = ["reqwest"] }
+
+[dev-dependencies]
+opentelemetry_sdk = { version = "0.31", features = ["rt-tokio", "trace", "testing"] }
+serial_test = "3"
+tokio = { version = "1", features = ["test-util"] }
