+++ sdk/packages/rust/helpers/src/lib.rs
@@ -1,6 +1,7 @@
 //! iii helpers.
 
 pub mod http;
+pub mod observability;
 pub mod queue;
 pub mod stream;
 pub mod worker_connection_manager;
