+++ sdk/packages/rust/helpers/src/observability/telemetry/otel_worker_gauges.rs
